Texture hotreloading

+++ LowCore/src/LowCoreUiWidgetInstance.cpp
@@ -98,7 +98,9 @@ namespace Low {
           Low::Util::HandleLock<WidgetInstance> l_Lock(get_id());
           // LOW_CODEGEN:BEGIN:CUSTOM:DESTROY
 
-          get_root().destroy_with_hierarchy();
+          if (get_root().is_alive()) {
+            get_root().destroy_with_hierarchy();
+          }
 
           // LOW_CODEGEN::END::CUSTOM:DESTROY
         }
